moved portainer fixture to fixtures folder
