+++ tests/conftest.py
@@ -15,6 +15,10 @@ import simcore_service_deployment_agent
 ## HELPERS
 current_dir = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
 
+
+## FIXTURES
+pytest_plugins = ["fixtures.fixture_portainer"]
+
 ## DIRs
